TLU: remove innecessary cast

+++ user/tlu/module/src/TluRawEvent2StdEventConverter.cc
@@ -31,7 +31,7 @@ bool TluRawEvent2StdEventConverter::Converting(eudaq::EventSPC d1, eudaq::Standa
   }
 
   // Set times for StdEvent in picoseconds (timestamps provided in nanoseconds):
-  d2->SetTimeBegin(static_cast<uint64_t>(d1->GetTimestampBegin()) * 1000);
-  d2->SetTimeEnd(static_cast<uint64_t>(d1->GetTimestampEnd()) * 1000);
+  d2->SetTimeBegin(d1->GetTimestampBegin() * 1000);
+  d2->SetTimeEnd(d1->GetTimestampEnd() * 1000);
   return true;
 }
